requirements.txt
streamlit==1.28.1
openai==1.3.5
langchain==0.0.335
langchain-openai==0.0.2
pydantic==2.5.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
python-dotenv==1.0.0
requests==2.31.0
